new modules trajectory/(parse|extract).ts split out from index.ts and utils.ts
